backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-generativeai==0.3.2
qdrant-client==1.7.0
python-dotenv==1.0.0
requests==2.31.0
nltk==3.8.1
beautifulsoup4==4.12.2
markdown==3.5.2
tiktoken==0.5.1